Move layout back from LoginWrapper to Layout.
This way we don't have to use the LoginWrapper on every page.
It was always the same, so no need to replicate it in the children.
